sharing code between both string cleaning and encoding binding
