+++ src/mlpack/methods/preprocess/preprocess_string_main.cpp
@@ -9,13 +9,7 @@
  * 3-clause BSD license along with mlpack.  If not, see
  * http://www.opensource.org/licenses/BSD-3-Clause for more information.
  */
-#include <mlpack/prereqs.hpp>
-#include <mlpack/core/util/mlpack_main.hpp>
-#include <mlpack/core/math/random.hpp>
-#include <mlpack/core/util/cli.hpp>
-#include <mlpack/core/data/string_cleaning.hpp>
-#include <mlpack/core/data/extension.hpp>
-#include <mlpack/core/data/tokenizer/split_by_char.hpp>
+#include "mlpack/methods/preprocess/preprocess_string_util.cpp"
 
 PROGRAM_INFO("preprocess_string",
     // Short description.
@@ -69,102 +63,6 @@ using namespace arma;
 using namespace std;
 
 /**
- * Function neccessary to create a vector<vector<string>> by readin
- * the contents of a file.
- *
- * @param filename Name of the file whose contents need to be preproccessed.
- * @param columnDelimiter Delimiter used to split the columns of file.
- */
-static vector<vector<string>> CreateDataset(const string& filename,
-                                            char columnDelimiter)
-{
-  vector<vector<string>> dataset;
-  // Extracting the Contents of file
-  // File stream.
-  ifstream fin(filename);
-  if (!fin.is_open())
-    Log::Fatal << "Unable to open input file" << endl;
-  string line, word;
-  while (getline(fin, line))
-  {
-    stringstream streamLine(line);
-    dataset.emplace_back();
-    while (getline(streamLine, word, columnDelimiter))
-    {
-      dataset.back().push_back(word);
-    }
-  }
-  return dataset;
-}
-
-static bool IsNumber(const string& column)
-{
-  for (auto i : column)
-    if (!isdigit(i))
-      return false;
-  return true;
-}
-
-/**
- * Function used to get the columns which has non numeric dataset.
- *
- * @param tempDimesnion A vector of string passed which has column number or
- *    column ranges.
- */
-static unordered_set<size_t> GetColumnIndices(const
-                                              vector<string>& tempDimension)
-{
-  unordered_set<size_t> dimensions;
-  int columnstartindex, columnendindex;
-  size_t found;
-  for (size_t i = 0; i < tempDimension.size(); i++)
-  {
-    found = tempDimension[i].find('-');
-    if ( found != string::npos)
-    {
-      try
-      {
-        // Has a range include, something of type a-b.
-        string subStringStart = tempDimension[i].substr(0, found);
-        string subStringEnd = tempDimension[i].substr(found + 1,
-            tempDimension[i].length());
-        if (!IsNumber(subStringStart) || !IsNumber(subStringEnd))
-        {
-          Log::Fatal << "Dimension value accepts only digits, characters"
-              " encountered, please recheck" << endl;
-        }
-        columnstartindex = stoi(subStringStart);
-        columnendindex = stoi(subStringEnd);
-        for (int i = columnstartindex; i <= columnendindex; i++)
-          dimensions.insert(i);
-      }
-      catch (const exception& e)
-      {
-        Log::Fatal << "Dimension value not clear, either negatve or can't "
-        "parse the range. Usage a-b" << endl;
-      }
-    }
-    else
-    {
-      try
-      {
-        if (!IsNumber(tempDimension[i]))
-        {
-          Log::Fatal << "Dimension value accepts only digits, characters"
-              " encountered, please recheck" << endl;
-        }
-        dimensions.insert(stoi(tempDimension[i]));
-      }
-      catch (const exception& e)
-      {
-        Log::Fatal << "Dimension value not appropriate " << endl;
-      }
-    }
-  }
-  return dimensions;
-}
-
-/**
  * Function used to write back the preproccessed data to a file.
  *
  * @param outputFilename Name of the file to save the data into.
@@ -211,7 +109,6 @@ static void mlpackMain()
   // Parse command line options.
   // Extracting the filename
   const string filename = CLI::GetParam<string>("actual_dataset");
-  // This is very dangerous, Let's add a check tommorrow.
   string columnDelimiter;
   if (CLI::HasParam("column_delimiter"))
   {
@@ -222,27 +119,7 @@ static void mlpackMain()
         "Delimiter should be either \\t (tab) or , (comma) or ' ' (space) ");
   }
   else
-  {
-      if (data::Extension(filename) == "csv")
-      {
-        columnDelimiter = ",";
-        Log::Warn << "Found csv Extension, taking , as "
-            "columnDelimiter." << endl;
-      }
-      else if (data::Extension(filename) == "tsv" ||
-          data::Extension(filename) == "txt")
-      {
-        columnDelimiter = "\t";
-        Log::Warn << "Found tsv or txt Extension, taking \\t as "
-        "columnDelimiter." << endl;
-      }
-      else
-      {
-        columnDelimiter = "\t";
-        Log::Warn << "columnDelimiter not specified, taking default"
-            " value" << endl;
-      }
-  }
+    columnDelimiter = ColumnDelimiterType(filename);
   // Handling Dimension vector
   vector<string> tempDimension =
       CLI::GetParam<vector<string> >("dimension");
